readme update and file cleanup
